pyproject.toml
[project]
name = "context-compactor"
version = "0.2.0"
description = "Automatic context window management for AI agent SDKs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = []  # Core has no required deps

[project.optional-dependencies]
# SDK adapters
pydantic-ai = ["pydantic-ai>=0.1.0"]
openai-agents = ["openai-agents>=0.0.3"]
claude-agent = ["claude-agent-sdk>=0.0.20"]
all-sdks = [
    "pydantic-ai>=0.1.0",
    "openai-agents>=0.0.3",
    "claude-agent-sdk>=0.0.20",
]

# Tokenizers
tiktoken = ["tiktoken>=0.7.0"]
litellm = ["litellm>=1.50.0"]

# Development
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

# Everything
full = [
    "context-compactor[all-sdks,tiktoken,litellm,dev]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["context_compactor"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
reportMissingImports = "warning"  # Optional deps may not be installed
reportInvalidTypeForm = "none"  # We use runtime type aliases intentionally

